Move MockUser class out of model sources
The MockUser class is a test class, so it shouldn't be part of the main model sources
The MockUser is only used in the build coordinator tests, so we can move it there
